Move VECTOR! datatype to extension
This does some cleaning up of VECTOR! and moving it to an extension.
One driving goal for this is to be able to make smaller builds, in
particular for inclusion as a JavaScript library.

While there are improvements, it is now a casualty of weaknesses in
the user defined type system, such that it cannot support enumeration
(via FOR-EACH/MAP-EACH/etc.)  However, that is something that needs
to be addressed with a general solution, if and when user-defined-types
are designed.