Add ZIP versioning and Unix file info to zip file
Several fields in unzip.reb were left zeroed out, such as for the
version of PKzip expected to be used, or the version the encoder was
aware of.  This also sets it to use Unix file permissions, and mirrors
the flag settings common to Unix Zip and MacOS Zip.

The intent is to fix a Mac OS shell-based unzipping problem, so it
hopefully resolves that, and perhaps some other things to.

Also brings unzip.reb up to date on the two-space semicolon convention
and uses `COMMENT <SECTION-NAME>` to mark unused sections.