GOB! to extension, eliminate HAS, revert MAKE, CONSTRUCT
This commit moves GOB! to an extension, which in order to implement
the derived GOB! MAKE functionality required getting rid of the
separation between MAKE and CONSTRUCT.

CONSTRUCT had been conceived as an arity-2 MAKE operation, to allow
for a spec and a body (similar to FUNCTION).  This never took off, and
MAKE PARENT-OBJECT SPEC has remained as the way of doing it.  Some
of the other reasons to avoid thinking in terms of MAKE OBJECT! were
cleaned up with generalized quoting.

So this puts CONSTRUCT back as arity-1.  However, it evaluates by
default--making it a candidate for replacing CONTEXT and OBJECT, as
a more verb-ish operator for the task.

The previous idea of using HAS as an arity-1 object maker as a parallel
to DOES is another thing that never took off, and the word is being
reclaimed to adapt to another purpose.