First step in moving EVENT! to an extension
This moves event proecssing code out of the core, while still
seeking to maintain the "single cell optimization" that was part
of the design goal of the original event abstraction.

The GC formalizes its feature of letting a cell indicate that its
extra field is a node and should be marked, as well as having
INFO flags in the REBSER node itself for saying that its LINK() and
MISC() fields require marking.  With this, there is now no knowledge
of events (specifically) in the GC.

To complete this requires the next commit--which transitions the
REBREQ to being a REBSER.  This way, the cell can be always telling
the truth that its extra field is a valid rebol node (or null),
because for the moment REBREQ* is just a random C structure.

Included in these changes are the ability to use any SYM_XXX legal
value as the "type" of an event.  It could be that a SYM_0 in this
spot could indicate that the event uses an expanded form, which
would have more room for full symbols (among other things).

It also brings VECTOR!, IMAGE!, and GOB! up to date with these new
changes to cells and series.