Move console port scheme into console extension
R3-Alpha tried to build a platform-independent port construct for
each port type (the p-xxx.c files) that would interact with OS-specific
code from the host.  This idea is being shifted in Ren-C to where
extensions basically implement platform specific ports however they
see fit.

This moves the p-console.c file into the console extension.  The next
step would be to take the service routines for implementing consoles
that are under the %os/ directory and merge those into the extension.
Then the extension could be built for a certain platform via #define
and use the correct files.