REBREQ => REBSER first step
To do I/O, R3-Alpha had the concept of "simple" devices, which would
represent abstractions of system services (Dev_Net would abstract the
network layer, Dev_File the filesystem, etc.)

There were a fixed list of commands these devices would handle (OPEN,
CONNECT, READ, WRITE, CLOSE, QUERY).  Further parameterization was done
with the fields of a specialized C structure called a REBREQ.

This layer was code solely used by Rebol, and needed access to data
resident in Rebol types.  For instance: if one is to ask to read from a
file, it makes sense to use Rebol's FILE!.  And if one is reading into
an existing BINARY! buffer, it makes sense to give the layer the
BINARY!.  But there was an uneasy situation of saying that these REBREQ
could not speak in Rebol types, resulting in things like picking
pointers out of the guts of Rebol cells and invoking unknown
interactions with the GC by putting them into a C struct.

This commit starts shifting to the idea where a REBREQ is actually a
REBARR, and able to hold full values (for starters, a REBSER*
containing binary data of what used to be in a REBREQ...which is
actually how PORT!s held a REBREQ in their state previously).  Then
the fields which it historically wanted the GC to process (next_req
and port_ctx) are now held in the LINK() and MISC() "custom node"
fields, which means the GC has no specific knowledge of what a REBREQ
is, but can still mark it.

!!! Due to the fact that "device" code does not have access to the
%sys-core.h API, doing this requires the addition of some transitional
libRebol APIs for working with REBREQ.  While they are inelegant, they
are much better at localizing the "dirtiness" than by having the
EVENT! layer need to hook the garbage collector directly--and can be
improved as time goes on.