Simplify %l-scan.c and update comments
The scanner has been reorganized as it has been understood.  One
aspect of this reorganization was to put the scanned token into the
scan state.  While this made sense at one time, it meant the code for
Locate_Token() couldn't simply `return TOKEN_XXX;` but had to do this
in two steps with `ss->token = TOKEN_XXX;` and then `return;`  Since
many of these statements were conditional and could have been single
lines, it made the already lengthy code lengthier.

This puts it back to returning the token.  While doing this, it also
updates comments and formatting in the file.