Simplify GROUP! invisible evaluation mechanic
The evaluation of groups had become somewhat complicated by the fact
that they wanted to be able to preserve the result in some cases,
e.g. `1 + 2 (comment "this should be 3)`...but did not have a spare
cell to evaluate into to save the prior result, because the GROUP!
itself might be resident in the spare cell.

These concerns don't seem to be relevant any longer.  Not only are
cells put into the feed's "lookback" if necessary, but the evaluation
routines are all obligated to not overwrite the input cell if there
are all invisibles or no content.  They can only mark the cell stale
if they did not modify the output.

The code is simplified.  Also, EVAL is changed to return void in the
case of invisibles (previously it asserted, because there was not clear
guidance on what it should do...but VOID! now has a fairly clear theory
which suggests it is perfect for such situations).