"CONST Compromise" - Loops and Function Bodies (extensible)
This implements the strategy that evaluation is mutable by default,
but that some functions take their arguments const, and propagate
that constness as a "wave" that must be overridden by MUTABLE.

Mechanically, this moves the const flag into a "feed flag", which
is either inherited or overridden.  Several calls into the evaluator
which would extract the array/index/specifier from a cell and pass
just that had to be modified to take the const/mutability bits of
the cell into account.

The `<const>` parameter flag is added.  Currently it has no effect on
RETURN values, but can/should be enhanced to apply constness there,
once that feature is articulated.