Revert Make_Arr() and Make_Ser() changes
It was experimentally tried to change Make_Array() and Make_Series()
and a few of their variants to be consistent with the REBARR and
REBSER types, and to help distinuish fro the ANY-ARRAY! and ANY-SERIES!
value types--which could then take the "full names".

This was only done on the creation functions for a start, to see how
it held up.  It was decided that it was ultimately bad--while perhaps
being more technically informative at the callsites for experts, it
compromised the readability to make the source more intimidating.  The
damage from intimidating source probably outweighs the technical
advantage for something that the compiler's type checking will handle
if you make a mistake.

https://forum.rebol.info/t/905/2?u=hostilefork