Have DO give explicit mutability to non-const input
The default behavior of inheriting the topmost frame's constness
unless there is a mutability override is a useful way of avoiding a
parameterization explosion, passing const or non-const to every API
that does an execution.  However, if you have a situation like:

    loop 2 [do block]

The DO wants to be driven by the block's constness, not by the wave
of constness coming from its frame feed.  DO has the luxury of having
a full REBVAL* cell argument to itself--which a RELVAL inside of a
SWITCH or CASE does not, hence it can augment that with the
CELL_FLAG_EXPLICITLY_MUTABLE in order to ask that a non-const value
not inherit the feed's constness.