Separate CALL into Windows and POSIX files
Originally CALL was described as a native in the core, but had its
implementation entirely outsourced to functions from the "host" which
had no knowledge of Rebol datatypes.  This "abstract OS" model is an
idea that Ren-C has deemed to not be beneficial, and seeks instead to
put these kinds of things into "extensions", which can both call OS
APIs and interact with Rebol values--either through the libRebol API
or the %sys-core.h API--while still not being part of the core.

However, the original implementation required abstractions like an
"OS independent character" type--as opposed to embracing the REBVAL*
of a TEXT! as serving this purpose for currency.  These have been
excised over time, but CALL still has some of them--and there are bugs
related to them remaining.

This begins an effort to turn the OS-specific implementations of CALL
to be self-contained, by putting them in their own files.