Eliminate automatic CR LF translation on Windows file write
It is the general prescription that Windows does not use UCS-2/UTF-16
when doing writes a redirection to a file or to a pipe.  This embraces
not only the philosophy of using UTF-8 there for text encodings, but
also to not use the CR/LF format.

This is consistent with a direction where CR LF is not an endorsed
format, and considered a "foreign codec"--as different as any other
character encoding, and not supported for text loading by default.
If users wish to have CR LF files, they will have to use ENLINE for
the moment.