Renovate CALL implementations, eliminate OSCHR
The original CALL implementation was based on calling out to a hostkit
routine that would do the work after parameter were unpacked into
"OSCHR" elements.  This eliminates that abstraction, and goes over
the code to eliminate unnecessary conversions.

In doing so, it discovered that redirected I/O in Windows is generally
agnostic about the format--meaning many of the multibyte/widechar
translations were actually not necessary.  This may indicate that a
better sharing of functionality between the routines is possible, now
that some of the fog has been cleared.

Adds some basic tests of the /INPUT and /OUTPUT functions for TEXT!,
BINARY!, and FILE!.