Patches so that CONSOLE can be used by Web demo
The console was set to use Isolate on the module, meaning it would
not see changes to LIB/PRINT or LIB/INPUT.  This relaxes that, as
it is enough to bind to lib to not be disrupted.  (Other isolate
usages should be reviewed now that some module issues have been fixed.)

Also fixes a bug for when CONSOLE was used without a /PROVOKE.

Changes callers of BROWSE to not print "opening a browser" message,
but leave that to the BROWSE implementation itself to worry about.