Make rebU() not quote its topmost level
Previously this would not print the message:

    REBVAL *printer = rebRun(":print");
    rebElide(rebU(printer), "{Message}");
    rebRelease(printer);

This is becuase while rebU() was setting the unevaluative feed flag
for any splices it did (e.g. `rebU("[", item1, item2, "]")`) the value
that got produced was not being marked unevaluative, hence it was
quote escaped like anything else.

Now it prints the message, as if you had said `rebEVAL, printer`.
This can be useful in particular if you are trying to fit the call
into one pointer slot (it became important in the JS API implementation
for rebPromise())