Move promise handling entirely into JavaScript extension
In order to implement JavaScript promises with pthreads, the
`<pthread.h>` dependency will be more invasive, and the JS-AWAITER will
require knowledge of the mutexes and pthread conditions currently in
effect.  Hence it's a good time to move away from the temporary
inclusion of `<emscripten.h>` in the main libRebol implementation, and
have the JavaScript extension do its work using native helpers.

This adds two new natives to assist in doing what the APIs did before:
REBPROMISE-HELPER and REBPROMISE-CALLBACK.  They are put into global
variables at rebStartup() time in JavaScript, and called where the
specialized APIs used to be.