First crack at implementing pure WASM + pthreads
This is a massive overhaul which introduces pthreads to the JavaScript
extension.  It does so while trying to minimize the differences between
the Emterpreter-based build and the Pthreads build (though ultimately
the emterpreter build should be phased out).

Stack overflow detection does not work yet, and failure handling is
not sorted out in general.  Cancellation is another thing that has to
be handled.  But it does enough to work with the web console.