Improve commenting in console extension
Given that the console code is currently linked to from the web REPL,
it should be somewhat acceptable to look at.  This updates the comments
and fixes a couple of issues.

It makes use of the new `===`-as-divider comment, which this now adds
into %base-defs.r for general availability.  The idea of using this
to skip over plain Rebol words in a MarkDown style is "edgy" as they
are not true comments and are variadic.  However, this puts the idea
out there for consideration (another consideration is there being a
mode where these are HIJACK'd and printed out, which could be very
useful).

Also takes away END from its role as a "tombstone" due to its potential
for accidents, renaming that to |||.