Console enfix trick, allow SHOVE -> into SET-WORD!/etc.
This makes it possible to use the SHOVE operator to assign a SET-WORD!
(or SET-PATH!, SET-GROUP!, SET-BLOCK!) from its left:

   >> 10 -> x:
   == 10

   >> x
   == 10

It also updates the console to inject its last result *quoted* as the
first item in the code being executed (unless it was void).  This
permits console users easy access to the last result, from the left.

    >> 1 + 2
    == 3

    >> * 10
    == 30

    >> -> x:
    == 30

    >> x
    == 30