Add Node.js support, change rebRun() => reb.Run() in JS
This makes some changes which allow libRebol to be run under Node.js

Because Node requires all exports to come from a module, the library
has to be loaded e.g. with:

    var reb = require('rebol')

All entry points must then be part of that module.  It would be
redundant to say `reb.rebRun()`, so this makes the actual exports not
have a prefix...thus it is `reb.Run()`.

While browsers do not have the same rule yet, a theoretical future
trend is that browsers will go the same way of modules as Node.  This
hasn't become mainstream yet.  However, it makes more sense to make
the browser match the Node model than to match the C model.  So in
the browser, one now says `reb.Run()` instead of `rebRun()` as well.

Users of the Node library would presumably need the same kind of
interaction with code asynchronously that the browser needs, to do
anything interesting.  There are also not currently build settings for
building without rebPromise() or the JS-AWAITER.  This means that
either WASM+PTHREAD or the emterpreter must be used...but Emscripten
does not yet build to node with WASM and PTHREAD.  Hence the
emterpreter must be used, making this a less practical solution at the
moment.  This is expected to change in the future.

Since emterpreter bytecode must be loaded for this to work, the build
script generates a --pre-js preloader which is put into the file.
At runtime, wherever the libr3.js is loaded from must have the bytecode
available to load.

There are many issues to consider, and rebPromise has not been worked
on yet for Node.  But this is a first step that can get the evaluator
to run basic code.