Move UTF-16 and UTF-32 codecs to extension
Ren-C's focus is on UTF-8, and there's not really a particularly strong
need for native-coded UTF-16 big-endian or little-endian conversions.
However, the code existed and was used as a test of the improved codec
mechanism.

This keeps the code, but moves it to an extension, so that core builds
(e.g. JavaScript library) can be smaller without it.