Switch `verb` parameter to `const`
R3-Alpha had an enumerated list of "action" ID numbers, which Ren-C
folded into the already existing list of "symbol" ID numbers for words.

But rather than pass a symbol number ID, a WORD! value was used--which
would encompass the idea of adding actions after-the-fact which did
not have pre-cooked symbol ID numbers.  This was passed as a non-const
value for some reason, so this changes that to pass it as const.