Simplify DROP_GC_GUARD
The DROP_GC_GUARD macro uses panic(), which in a valgrind or address
sanitizer build will report both the current call stack and the stack
where a node was allocated.  Considering how rarely this fires (and
how rarely the stack wouldn't point to the smoking gun of what the
problem is immediately), it doesn't need an added layer of complexity
to try and capture the file and line where it is called via C macros.