Make THEN not voidify result, but give branch even if null
Previously, THEN would not pass through its branch result directly if
it were null.  So in the following case:

    if condition [...] then [...] else [...]

There would have been no way that the ELSE would run if the condition
were true.  THEN would be forced to void at least.

Such a tying to the condition would make sense in a language which
hinged on THEN for its true branch, e.g.:

    if condition then [...] else [...]

However, the nature of THEN and ELSE is such that they are more useful
constructs in a general chaining sense if they act on the clause to
their immediate left, and leave the voidifying to the conditional
expression (IF, SWITCH, CASE).  This means you can use:

    if condition then [...] then [...] then [...]

And it can meaningfully act on each step's result, rather than mandate
that all of them run because one of them did.

The simple mitigation for anyone wanting a THEN [...] ELSE [...] to
act in a way that guarantees the ELSE and THEN are mutually exclusive
is to ensure that the THEN returns a non-NULL value manually.  If the
result of the THEN is not important, one can use ALSO [...] ELSE [...]
because that will not carry through the ALSO's branch at all, merely
telegraph the result that cued it to run (or the null that caused it
to not run)