Make StdIO and console port READ go direct to BINARY!
The R3-Alpha device model was based on building a layer of code that
was not aware of Rebol types or values.  However, these devices were
typically trying to build Rebol values.  This led to the creation of
a fairly superfluous layer of C structures for communicating Rebol
concepts to an "abstract OS".

Another downside of this is that some things like BINARY! buffers were
being handed to this layer as raw C pointers.  This meant that updates
were being performed on these series without keeping other rules
in sync (such that the rule that BINARY! series must be terminated
with a 0 byte...a rule that has been kept in effect, in order to
enable the same memory backing a binary to be used to back a TEXT!)
Ultimately the series would still need to be terminated, but a GC or
other intermediate step could wind up seeing the unterminated form.

This changes the console reading code that is behind INPUT to be
aware that it is dealing with a BINARY!.  This is in the spirit of
continuing to simplify the code to a more manageable level.