Use BINARY! direct value for network device layer input
The device and port interactions were quite snakey, and one aspect of
that was that a raw byte pointer into a BINARY! value was passed to
the layer to be READ into.  Updates to that buffer would happen outside
of the sanctioned binary-update methods which synchronize size and
0 byte termination.  (BINARY! in R3-alpha had a zero byte terminator
in order to parallel ASCII strings, UTF-8 everywhere keeps this to make
it easier to view binaries as strings in the same buffer--a concept
which is in the works).

Because these modifications were happening unknown to the series layer,
asserts would trigger.  Rather than get rid of the asserts, this
pushes the device code to do a legitimate binary update, and tries to
get a handle on other aspects of the port/device protocol with asserts.
This is enough to get `read https://example.com` to run without
an assert.  (The goal is to replace/rewrite the affected code, so the
main thing is just having enough assertion of invariants to improve it,
vs. fixing it completely in one commit.)