Don't use CELL_FLAG_FALSEY in PG_R_Redo, cleanup
In the goal of reclaiming CELL_FLAG_FALSEY, this switches the redo
instruction to use the extra field in the "pseudotype" cell instead.

It also gets rid of a somewhat obfuscating safety precaution that was
taken with global cells to avoid incrementing one cell and reaching
another.  Not only has this particular bug never been hit, but address
sanitizer probably already takes care of this by inserting poisoned
bytes between globals or stack locals to see when you've stepped out
of one and into the next.