Reclaim CELL_FLAG_FALSEY
The "falsey" flag was an early optimization when cell flags were more
available for random usages.  The idea was that whether something was
truthy or falsey didn't have to go through a type test but would be
cached in a flag--hence LOGIC! had no need for a payload, the header
flag was enough.

Various changes have made cell flags scarce, and it's possible to do a
pretty good optimization just by organizing the type order better.
This frees the flag for more important applications.