Change custom node usage to payload first, not EXTRA()
This is based on planning to share more generic facilities based on
using tests of node flags...e.g. common tests for mutability between
an ANY-CONTEXT! and an ANY-SERIES!.

Since this will be used for core types now and not just custom ones,
rename the paylaod extra to "Any".  This means shorter code, since
`PAYLOAD(Any, XXX)` is briefer than `PAYLOAD(Custom, XXX)`.