Remove BITSET! from ANY-SERIES!, related fixes
There was overlap in implementation between BITSET! and BINARY!, which
led it to be in the ANY_SERIES() class and initialized with
Init_Any_Series().  Yet it did not have an index or behave like a
series type.

With a generic "FIRST_IS_NODE" payload strategy, bitsets can leverage
shared code with series without needing to be in the category.

This also gets rid of REMOVE/MAP, since setting a key to null does a
removal.