Fix REWORD behavior with 0-arity functions
REWORD was passing the matched keyword as an argument to a substitution
function, but if the function took no arguments it was mistaking the
unused argument for the function's result.

This fixes that, and also tightens up REWORD's source so it does not
have so much whitespace/commenting.

Breaks reword tests out into their own file.