Make ANY-STRING! and BINARY! types unbindable
Moving values in the system is tied up with "bindability", and knowing
if something is bindable or not has to be a fast test.  Whether a type
is bindable or not is tested by its order in the fundamental types:
all types below a certain index are bindable, and above it are not.

Types that are not bindable are free to use their "extra" field in
the cell for anything they want.  But types that are bindable have
the binding automatically managed.

Strings and binaries were not bindable, but in order to make the
ANY_SERIES() test fast and keep the ANY_INERT() test fast, they lost
use of their extra bits in the cell to a null binding.  This limits
the application of that slot for other purposes.  in UTF-8 everyehwere
it is considered that the slot would be used as a per-value cache/stamp
to accelerate indexing.