Reorder types to put NULLED/VOID/BLANK/LOGIC at beginning
This changes the order in types.r to put all the truthy types above
REB_LOGIC.  It opens the doors to further optimizations, though these
would require allowing pseudotypes to live in the middle of the
range of valid types--which is not that unthinkable.

For easier editing, this moves the macros that relate to the type
ordering to %sys-ordered.h.