Inert Optimization
It seems like it should be pretty simple to fulfill the arguments of
ADD for `add 1 2` without too much overhead.  They are inert values,
and there should be a cheap test for that.  (There is: ANY_INERT is
a single greater than comparison on a byte.)

The main problem comes from the somewhat detailed logic regarding how
enfix is handled--which has a large number of edge cases.  That logic
plays out in Eval_Core_Throws() and one doesn't want to repeat it.

This takes a small and conservative step toward optimization without
endangering too much behavior.  It repeats some amount of code, but
is geared in a way that it could likely be factored out of the
evaluator better as the optimization improves.