Eval_Core() simplifications
Evaluations were clearing off the VALUE_FLAG_UNEVALUATED bit in cases
which were not argument fulfillments.  However, since the bit is not
heeded anywhere in the system *but* argument fulfillment, it's not
particularly necessary to lose the information if not.  So this takes
away that clearing.

Returning a thrown bit from Eval_Core() could be done more simply
without a local variable, so this eliminates the local variable.

As a step toward simplifying FEED_FLAG_NO_LOOKAHEAD, this takes some
obfuscating uses that went through an indirect routine and just works
with the flag directly.