Eliminate EVAL_FLAG_TO_END
In the early implementation of "frames" in the evaluator, it was too
expensive to create a frame for every evaluative step.  To optimize
this, the evaluator could run in two modes...one that did a single
step, and another that ran an array to the end.  This also allowed it
to do things like reuse the work of a variable fetch for lookahead
that did not turn out to be enfix, to get as an ordinary variable.

With time, frames and feeds were factored to be easier to make and
reuse.  Also, the Eval_Core() routine aims to be hookable...such that
it could be hooked after any step...and that would mean that Eval_Core
could not take for granted that it was supposed to be running the
same evaluator function.

This eliminates the TO_END flag, which starts a wave of simplification.
(and also reclaims an evaluator flag)