Pare down Do/Eval/Array_At Core to two variations
There were three variants of evaluating a REBVAL/RELVAL fully.  Now
that "eval" is always single-stepped, it makes sense to call the
multi-step variations "DO".  Also, having a variation that takes a
specified value isn't all that much of an advantage...clients of the
internal API know about specifiers.  So this uses the core distinction
for allowing the specification of a preloaded value or flags.