Make frame flags guaranteed to be restored to originals
This changes the behavior of Eval_Core_Throws() to preserve the
original flags (unless they are consumed by the evaluator).  This means
calling in a loop need not worry about the flags becoming corrupt
between evaluations.

As a consequence, Eval_Step_Mid_Frame_Throws() is no longer needed,
since all it did was save and restore frame flags.