Tighten handling of OUT_MARKED_STALE flag
Several features now depend on the ability of the evaluator to preload
a value into the output cell, and then have that be the value if no
evaluations are performed.  So example, the reason that `do []` is
void is that it runs Init_Void() on the output cell, and then nothing
overwrites that.  The same is true of `do [comment "hi"]`.

This requires some management of the left behind value, so it is not
inadvertently seen as a candidate for enfix.  So the value is marked
to be "stale" using a cell flag.  What's good about making it a cell
flag is that initializations with new values overwites it implicitly...
which is exactly when one would want to see it cleared.

But due to its informational nature to the evaluator, the flag has to
be tightly controlled.  e.g. when a `do []` is used to produce the
result of a native, that result can't have OUT_MARKED_STALE on it or
it might confuse `do [] else [...]` into acting like `(else [...])`

Early on there was a frame flag registered for interest in preserving
the stale flag state.  Then that flag was eliminated, and it was
considered an internal evaluator only feature to discern the stale
flag.  It turns out there are too many compelling cases for being able
to sense the flag, so this instead makes an Eval_Throws() inline
function that wraps the idea of calling the hookable evaluator and
then clearing off the flag.  Then, any functions that still want to
deal with it access it through the menacing name of
"Eval_Core_Maybe_Stale_Throws()".

Includes related cleanups which help make it clear which routines fully
overwrite their outputs vs `Maybe_Stale`.