Delete repetition of Apache License no warranty clause
The comment headers for files in the project are already rather
lengthy.  Picking a long paragraph out of the license to quote verbatim
in each and every file is rather pointless--considering that it's in
the license, and there's a link to that license, and saying that you
can only use it under that license.  It just means it takes that much
longer to scroll down to get to real code and comments.