Debug build speedup: ASSERT_NOT_END()
Due to the debug build not inlining things, some common functions that
get called disproportiately many times wind up showing up on the radar
of the debug build runtime...when they're only supposed to be single
instructions in the release build.

This mitigates a situation with assert(NOT_END()), which was doing not
only a check that the kind byte wasn't an end but that the cell was
otherwise in good shape and not corrupt.  That seems excessive because
such tests would be done by whatever the code usually would do with
the value once it stated it's not an end.  Hence it's enough to make
it a macro with no debug build function call overhead and let other
code pick up the slack on the deeper checks.