Faster checking for bar ('|'), faster stack debug check
There's no longer a BAR! datatype.  But checking for BAR was costing
more than it needed to.  Since it doesn't have case, there's no need
to do canonization or dereference the string node to get the symbol,
it can just be done as a direct comparison against a global.

That helps in both the debug and release build.  But this also helps
the debug build by turning some inline functions (that don't get
inlined for debugging) into macros, such as ones used in doing a
check of a frame's ->out pointer to make sure it's not an argument
to any other frame that could be made visible.