"evil macro" speedups for non-inlined debug build
inline functions should be used for safety whenever anything repeats
the evaluation of an argument.  However, since the debug build does
not inline functions, one can wind up with primitives that are called
extremely often that are built up of smaller primitives that don't
inline...and the result can mean trivial operations take a significant
amount of runtime when running with debug checks.

This uses macros in a strategic way to throw in a few speedups to some
extremely common operations (like VAL_INDEX() or VAL_SERIES()) so that
they can continue to check safely, at the cost of a slight bit of
obfuscation.  But in this case it's worth it.