Reorganize %sys-eval.h and %sys-do.h
A separation that "eval(uate) means single step" and "do means multiple
steps" was started a while ago.  It also suggested that DO would not
be able to remember any results prior to the DO, while EVAL would have
that capability by the means of "stale" values.

This separation of responsibility has continued to concretize, with
there no longer being a form of "evaluate to end"--instead pursuing
other optimizations.  This does some straightening of the code and
reducing the overall redundancy.