Eliminate "Reb_Indexor" data type
R3-Alpha merged together several return concepts into an unsigned
integer, that could be either an index position OR a flag indicating
a throw OR a flag indicating one had come to the end of an evaluation.

This conflation made it easy to get wrong and not check for the flags
properly, or to pass flags when you meant to pass an index.  Ren-C
chipped away at this problem over time...but one of the early steps
was to have a special "REBIXO" datatype to represent an "index OR a
flag".  This type had some special compile-time and run-time
protections, which would alert you if you tried to pass a flag that
was signaling a throw or an end to a place that was expecting a
plain array index.

However, better techniques now exist for handling the problem.  Most
all interaction with evaluation is done through the frame/feed, where
any concept of indexing is abstracted (whether it is a va_list or
an array, or hybridizations of those things with reevaluated values
spliced in--with more plans for this to expand in the future).

And the only real case where a frame and feed get torn down and then
reconstructed via index is EVALUATE (the parallel to old DO/NEXT).
Basically nothing else uses it.  The index can be handled as an
output parameter by pointer, which can be inlined...and the thrown
flag as a boolean return result, also inlined.  Signaling end can
now be done by means of END cells.  This division gives a greater
natural protection against errors, and is self-documenting.

(PARSE was converted to use REBIXO so that the thrown flag would be
returned in addition to indexes and NOT_FOUND vs. using a thrown bit
on the output cell itself.  At first this had been commentary, but
is now required since there is no thrown bit intrinsic to values and
it's *only* conveyed by stack return results.  This reduces that to
just using a REBCNT for now, but it should be converted to the bool
returning convention.)