Reduce and focus comments in %c-eval.c
This optimizes the `f->special` checks and makes their meaning more
clear in the source.  Also takes some more detailed comments and moves
them to %d-eval.c, while moving the most important "big picture"
remarks to the header of %c-eval.c

Also substitutes _Core_ with the more menacing _Internal_, as many
functions with "Core" in the name referred to being a version of the
function with more options--while still calling the hooked version of
the evaluator.