Make SET-GROUP! and GET-GROUP! setify and getify in COMPOSE
The first behavior chosen for how arbitrary quotes would be used in
COMPOSE was that they would protect GROUP!s from one level of
evaluation, thus allowing a way to "escape" groups in COMPOSE:

    >> compose [(1 + 2) '(1 + 2) ''(1 + 2)]
    == [3 (1 + 2) '(1 + 2)]

The introduction of tagged/labeled COMPOSE gave a more attractive way
to point out just the GROUP!s you wanted to affect.  Also, the
general usefulness of being able to compose and produce a quoted value
made the alternate interpretation more sensible.  This changes it to:

    >> compose [(1 + 2) '(1 + 2) ''(1 + 2)]
    == [3 '3 ''3]

In a similar spirit, the introduction of GET-GROUP! and SET-GROUP!
offer an ability to do something interesting.  This makes that to be
to turn whatever the evaluated product is into its SET-XXX! or GET-XXX!
representation:

    >> var: 'foo
    >> compose [(var): 1 + 2]
    == [foo: 1 + 2]

    >> var: '(foo bar)
    >> compose [(var): 1 + 2]
    == [(foo bar): 1 + 2]

If it is an ANY-STRING! it is turned into an unbound word.  An
experiment runs the code for cases that don't have an obvious transfrom
through TO SET-WORD! and TO GET-WORD!

    >> var: "foo"
    >> compose [(var): 1 + 2]
    == [foo: 1 + 2]  ; unbound SET-WORD!

!!! This commit raises a question about COMPOSE's splicing-by-default
behavior.  It's not clear what to do with something like '(var) or
(var): when it is a BLOCK! and being spliced.  This suggests that
perhaps the (( )) trick should be used to indicate splicing in a
COMPOSE, and that all non-doubled groups do not splice.