Change rebTextW/etc. API names to rebTextWide/etc.
The choice to make the APIs for WCHAR end in plain "W" was based on
trying to match Windows's convention.  However, this isn't necessarily
clear, and it creates a bit of a problem with the pervasive Q suffix
being added so it says "rebTextWQ", which is one too many single-letter
suffix.  Since the -W APIs are not particularly favored (UTF-8 is the
main target for the system, and WCHARs are on their way out), there's
no reason in particular to give it a short name.