Plan rebSpellInto()/rebBytesInto() for future variadic
Generally all of the libRebol extractors are variadic.  However, the API
wrapping code which makes wrappers for JavaScript doesn't cover the
case where *all* the arguments aren't variadic, so the "output"
pointers break that.

This should be a temporary situation, so this "futureproofs" it by
making sure the C89-oriented code that calls it passes a rebEND as
a parameter.