Add "Q" variations of APIs (e.g. rebQ, rebSpellQ)
There was significant deliberation over what the following code does:

    REBVAL *word = rebRun("'print");
    REBVAL *type = rebRun("type of", word);

If the WORD! is simply spliced into the code and run, then that will be
an error.  It would be as if you had written:

    do compose [type of (word)]

It may seem to be more desirable to pretend you had fetched word from a
variable, as if the code had been Rebol.  The illusion could be given
by automatically splicing quotes, but doing this without being asked
creates other negative side effects:

     REBVAL *x = rebInteger(10);
     REBVAL *y = rebInteger(20);
     REBVAL *coordinate = rebRun("[", x, y, "]");

You don't want to wind up with `['10 '20]` in that block.  So automatic
splicing with quotes is fraught with problems.  But it might be useful
sometimes, so it is exposed via `rebRunQ()` and other `rebXxxQ()`.

These facilities are generalized so that one may add and drop quoting
from splices on a feed via ranges, countering any additions via rebQ()
with a corresponding rebU().